remove question about int/float coverage profiles. Usually we need integer profiles (for pasio segmentation), but float is more universal. Now it's possible to explicitly pass dtype if necessary - and we use it

+++ extract_profile.py
@@ -5,7 +5,7 @@ from coverage_profile import transcript_coverages_from_file
 cds_annotation_fn = sys.argv[1] # 'gencode.vM22.cds_features.tsv'
 alignment_filename = sys.argv[2] # 'METTL3_ribo_Coots2017_m_r2.bam'
 cds_info_by_transcript = CodingTranscriptInfo.load_transcript_cds_info(cds_annotation_fn)
-for (transcript_id, profile) in transcript_coverages_from_file(alignment_filename, cds_info_by_transcript):
+for (transcript_id, profile) in transcript_coverages_from_file(alignment_filename, cds_info_by_transcript, dtype=int):
     if transcript_id in cds_info_by_transcript:
         transcript_info = cds_info_by_transcript[transcript_id]
         cds_profile = transcript_info.cds_profile(profile)
